fix: give the escape-hex round-trip its own class and tag
